Move image cache implementation to the net crate
