+++ components/constellation/lib.rs
@@ -24,6 +24,7 @@ extern crate layout_traits;
 #[macro_use]
 extern crate log;
 extern crate msg;
+extern crate net;
 extern crate net_traits;
 extern crate offscreen_gl_context;
 extern crate profile_traits;
