+++ components/net/lib.rs
@@ -4,6 +4,7 @@
 
 #![deny(unsafe_code)]
 #![feature(box_syntax)]
+#![feature(step_by)]
 
 extern crate base64;
 extern crate brotli;
@@ -12,6 +13,7 @@ extern crate devtools_traits;
 extern crate flate2;
 extern crate hyper;
 extern crate hyper_serde;
+extern crate immeta;
 extern crate ipc_channel;
 #[macro_use] extern crate log;
 #[macro_use] #[no_link] extern crate matches;
@@ -35,6 +37,7 @@ extern crate tinyfiledialogs;
 extern crate unicase;
 extern crate url;
 extern crate uuid;
+extern crate webrender_traits;
 extern crate websocket;
 
 mod blob_loader;
@@ -46,6 +49,7 @@ mod data_loader;
 pub mod filemanager_thread;
 pub mod hsts;
 mod http_loader;
+pub mod image_cache;
 pub mod mime_classifier;
 pub mod resource_thread;
 mod storage_thread;
